Updated community settings: added hook form, yup, validations
