+++ package.json
@@ -27,6 +27,7 @@
     "@floating-ui/react-dom": "^2.1.2",
     "@hiveio/dhive": "^1.3.1-beta",
     "@hiveio/hivescript": "^1.3.3",
+    "@hookform/resolvers": "^5.2.0",
     "@joplin/turndown-plugin-gfm": "^1.0.62",
     "@react-spring/web": "^9.7.3",
     "@sentry/browser": "^8.33.1",
@@ -91,6 +92,7 @@
     "react-google-recaptcha": "^3.1.0",
     "react-grid-layout": "^1.4.4",
     "react-highcharts": "^16.1.0",
+    "react-hook-form": "^7.61.1",
     "react-in-viewport": "1.0.0-beta.6",
     "react-resizable": "^3.0.5",
     "react-resize-detector": "^7.1.2",
