--- src/app/(dynamicPages)/community/[community]/_components/community-settings/_index.scss
@@ -1,15 +1,0 @@
-@import "../../../../../../styles/vars_mixins";
-
-.community-settings-dialog {
-  .community-settings-dialog-content {
-    .settings-form {
-      padding-top: 10px;
-
-      textarea {
-        border-radius: 6px;
-        height: 120px;
-        resize: none;
-      }
-    }
-  }
-}
