+++ src/features/ui/input/form-controls/index.tsx
@@ -1,28 +1,34 @@
-import React, { MutableRefObject } from "react";
+import React, { forwardRef } from "react";
 import { Textarea, TextareaProps } from "./textarea";
 import { Select, SelectProps } from "./select";
 import { Input, InputProps } from "./input";
 import { Checkbox, CheckboxProps } from "./checkbox";
 import { Toggle } from "@/features/ui/input/form-controls/toggle";
 
-type Props = (InputProps | TextareaProps | SelectProps | CheckboxProps) & {
-  ref?: MutableRefObject<any>;
-};
+type Props = InputProps | TextareaProps | SelectProps | CheckboxProps;
 
-export function FormControl(props: Props) {
+type Refs = typeof Textarea | typeof Select | typeof Toggle | typeof Checkbox | typeof Input;
+
+export const FormControl = forwardRef<Refs, Props>((props, ref) => {
   switch (props.type) {
     case "textarea":
-      return <Textarea {...props} />;
+      return <Textarea {...props} ref={ref as any} />;
     case "select":
-      return <Select {...props}>{props.children}</Select>;
+      return (
+        <Select {...props} ref={ref as any}>
+          {props.children}
+        </Select>
+      );
     case "checkbox":
       if (props.isToggle) {
-        return <Toggle {...props} />;
+        return <Toggle {...props} ref={ref as any} />;
       }
-      return <Checkbox {...props} />;
+      return <Checkbox {...props} ref={ref as any} />;
     default:
-      return <Input {...props} />;
+      return <Input {...props} ref={ref as any} />;
   }
-}
+});
+
+FormControl.displayName = "FormControl";
 
 export * from "./input-skeleton-loader";
